requirements.txt
python-slugify==5.0.0
pyqrcode
pypng
Pillow